init commit

+++ photo-settings-service/src/main/java/com/kideya/photosettingsservice/repository/SettingsRepository.java
@@ -2,9 +2,8 @@ package com.kideya.photosettingsservice.repository;
 
 import com.kideya.photosettingsservice.model.Settings;
 import org.springframework.data.mongodb.repository.MongoRepository;
-//import org.springframework.data.repository.CrudRepository;
 
 
 public interface SettingsRepository extends MongoRepository<Settings, Integer> {
-    Settings findByUserId(int userId);
+    Settings findByUserId(long userId);
 }
